Provide an option in discoverPeers to control the delay until the discoverPeers method on wifi manager is called.
Discover peers serves two purposes. It is used to kick Android to broadcast local services. It is also used for purposes of discovering wifi direct peers. In the case of kicking Android's processes, we want to wait for an interval period until actually calling the method. In the case of actually wanting to discover peers, it should start as soon as possible.